core: move EnumAttribute out of ir core
