+++ xdsl/dialects/linalg/attrs.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from enum import auto
 
-from xdsl.ir import EnumAttribute
+from xdsl.dialects.utils import EnumAttribute
 from xdsl.irdl import irdl_attr_definition
 from xdsl.parser import AttrParser
 from xdsl.printer import Printer
